Implement BinaryEncodable
The encoding was verified by decoding a .binarycookies file, then encoding it and writing to a file and diffing the files with colordiff and xxd.

There are some changes that came along with the addition of this conformance in order to get the output matching, like comments and new Flags values.

Speaking of those flags values, there are two that are now codable but the meaning is unknown. I'm assuming that one is the discard flag, but I'm not sure what the other one is or which is which. I'll need to do more testing, because so far I've just been using the main Cookies.binarycookies file on my computer to verify.

This also implements the checksum, which is the sum of every 4th byte in all of the pages.

Addresses #1.